fix : PlayerHistoryRow posseision -> posessission 으로 백엔드 통일 사항 반영했습니다
