--- src/1_Page/Championship/ui/DashBoard/ui/MatchLineupContainer/ui/VerticalPersonStatCards/ui/TeamTable/ui/PlayerHistoryRow/ui/PlayerDetailRow/type.d.ts
@@ -1,4 +1,0 @@
-type PlayerDetailRowProps = {
-  player: PlayerStats;
-  isMine?: boolean;
-};
