+++ src/1_Page/Championship/ui/DashBoard/ui/MatchLineupContainer/ui/VerticalPersonStatCards/ui/TeamTable/ui/PlayerHistoryRow/ui/StatProgressBar/index.tsx
